requirements.txt
Pillow==8.1.1
PyQt5==5.13.1
PyQt5-sip==12.7.0